recompose `_tool_suggest.py` and implement deps for remote mode
