+++ experiments/mcp-exps/src/agents/__init__.py
@@ -6,6 +6,7 @@ from ._tool_suggest import (
     TSAgentState,
     create_jsonl_repo_tool_suggest_deps,
     create_phase_scoped_tool_suggest_deps,
+    create_remote_phase_scoped_tool_suggest_deps,
     create_tool_suggest_agent,
     tool_suggest_run_result_processor,
 )
@@ -18,6 +19,7 @@ __all__ = [
     "create_basic_deps_maker",
     "create_jsonl_repo_tool_suggest_deps",
     "create_phase_scoped_tool_suggest_deps",
+    "create_remote_phase_scoped_tool_suggest_deps",
     "create_tool_suggest_agent",
     "tool_suggest_run_result_processor",
 ]
