+++ experiments/mcp-exps/src/agents/_tool_suggest/constants.py
@@ -1,0 +1,9 @@
+"""Constants shared by recording and JSONL replay."""
+
+IGNORED_TOOL_LABELS: frozenset[str] = frozenset(
+    {
+        # pydantic-ai ToolOutput names (final structured output)
+        "final_result",
+        "finish",
+    }
+)
