+++ experiments/mcp-exps/src/agents/_tool_suggest/phase_names.py
@@ -1,0 +1,8 @@
+"""Phase / collection naming for tool-suggest storage."""
+
+import re
+
+
+def sanitize_phase_name(name: str) -> str:
+    """Make phase name safe for filenames and remote collection names."""
+    return re.sub(r"[^\w\-]", "_", name) or "phase"
